% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs.R
\name{component-methods}
\alias{component-methods}
\title{Component Methods}
\description{
Methods are a list of functions that are contained within a given component,
and typically will do work with component data.

Methods can be accessed in both other methods and the component template
via the \code{self} keyword.
}
\details{
The following is an excerpt from a simple counter component:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{methods = list(
    setup = function(input, output, session) \{
        output$txt <- renderText(\{
            paste0("Count is: ", self$count())
        \})
        observeEvent(input$button, \{
            self$count(self$count() + 1L)
        \})
    \},
    increment = function() \{
        self$count(self$count() + 1L)
    \}
)
}\if{html}{\out{</div>}}
\subsection{Hooks}{
\subsection{Setup}{

The \code{setup} hook is defined by passing a function named "setup" to the
methods list. Setup is used as the module server for the component.

\if{html}{\out{<div class="sourceCode r">}}\preformatted{setup = function(input, output, session) \{

\}
}\if{html}{\out{</div>}}
}

\subsection{Render}{

The \code{render} hook is defined by passing a function named "render" to
the methods list. The render hook allows for the modification of template code
prior to its rendering.

\if{html}{\out{<div class="sourceCode r">}}\preformatted{render = function(element) \{

\}
}\if{html}{\out{</div>}}
}

}
}
\seealso{
Other components: 
\code{\link{component-data}},
\code{\link{component-styles}},
\code{\link{component-template}},
\code{\link{component}()}
}
\concept{components}
