% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component-styles.R
\name{component-styles}
\alias{component-styles}
\title{Component Styles}
\description{
The styles argument is a character vector of length 1 that defines the styles for the component.
Styles are scoped to the component.
}
\details{
The styles argument takes a character vector of length 1 that defines the styles for the component.
For example:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{styles = "a \{ color: red; \}"
}\if{html}{\out{</div>}}

This would define the CSS styles for the anchor elements (\verb{<a>}) in the component, setting their color to red.
Styles defined in this way are scoped to the component, meaning they will only apply to
elements within that component.
}
\seealso{
Other components: 
\code{\link{component-data}},
\code{\link{component-methods}},
\code{\link{component-template}},
\code{\link{component}()}
}
\concept{components}
