% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompose.R
\name{decompose}
\alias{decompose}
\title{Decompose a component instance}
\usage{
decompose(x)
}
\arguments{
\item{x}{a shiny tag returned from calling a component}
}
\value{
list
}
\description{
Given a component instance tag \code{x}, decompose the instance
into separate server and UI elements
}
