% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsx-package.R
\docType{package}
\name{rsx}
\alias{rsx-package}
\title{rsx: shiny components}
\description{
Create complex yet manageable Shiny applications with nestable components that encapsulate both UI and state. Heavily inspired by many JavaScript frameworks, particularly Vue.
}
\details{
The \strong{rsx} package implements encapsulated shiny module server &
UI combinations called "components".

Main exported methods:
\itemize{
\item making components: \code{\link[=component]{component()}}
\item manipulating components: \code{\link[=decompose]{decompose()}}
\item loading apps: \code{\link[=rsx_app]{rsx_app()}}, \code{\link[=rsx_server]{rsx_server()}}, \code{\link[=rsx_ui]{rsx_ui()}}
}

Website: \url{https://elianhugh.github.io/rsx/}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://elianhugh.github.io/rsx/}
}

}
\author{
\strong{Maintainer}: Elian Thiele-Evans \email{elianhte@gmail.com} (\href{https://orcid.org/0000-0001-8008-3165}{ORCID})

}
\keyword{internal}
