% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compilation.R
\name{rsx_module_server}
\alias{rsx_module_server}
\title{Create an rsx module server}
\usage{
rsx_module_server(id)
}
\arguments{
\item{id}{unique namespace}
}
\description{
\code{rsx_module_server} is a low-level function that loads up the module servers of all instantiated components.
This is similar to \code{rsx_server}, but also allows for namespacing
the rsx server.
}
