% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compilation.R
\name{rsx_ui}
\alias{rsx_ui}
\title{Create an rsx UI taglist}
\usage{
rsx_ui(root, id = NULL, app_class = "App", resource_path = NULL)
}
\arguments{
\item{root}{a component used as the top-level node for the shiny app}

\item{id}{TODO}

\item{app_class}{the html class attribute for the app wrapper}

\item{resource_path}{path to a resource folder, if \code{NULL} styles will be inlined}
}
\value{
\code{shiny::tagList} object
}
\description{
\code{rsx_ui} is a low-level function for creating rsx ui without the use of \code{rsx_app}.
}
\seealso{
Other compilation: 
\code{\link{rsx_app}()},
\code{\link{rsx_server}()}
}
\concept{compilation}
