% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs.R
\name{component-styles}
\alias{component-styles}
\title{Component Styles}
\description{
The styles argument is a named list of scoped and global CSS styles.
}
\details{
\if{html}{\out{<div class="sourceCode r">}}\preformatted{list(
 scoped = "",
 global = ""
)
}\if{html}{\out{</div>}}
}
\seealso{
Other components: 
\code{\link{component-data}},
\code{\link{component-methods}},
\code{\link{component-template}},
\code{\link{component}()}
}
\concept{components}
