% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component.R
\name{component}
\alias{component}
\title{Create a component}
\usage{
component(
  name = paste0("unnamed_component-", random_id()),
  data = NULL,
  methods = NULL,
  template = NULL,
  styles = NULL
)
}
\arguments{
\item{name}{TODO}

\item{data}{function that returns a named list}

\item{methods}{named list of functions}

\item{template}{function that returns a taglist}

\item{styles}{named list of "global" and "scoped" styles}
}
\description{
TODO
}
\seealso{
Other components: 
\code{\link{component-data}},
\code{\link{component-methods}},
\code{\link{component-styles}},
\code{\link{component-template}}
}
\concept{components}
