% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compilation.R
\name{rsx_server}
\alias{rsx_server}
\title{Create an rsx server}
\usage{
rsx_server()
}
\value{
shiny server function
}
\description{
\code{rsx_server} is a low-level function that loads up the module servers of all instantiated rsx::components.
If the \code{rsx_app} function is not viable for your shiny application setup, the rsx_server function
can be used.
}
\details{
Some things to note:
\itemize{
\item Due to randomised hashing of component namepaces, shiny modules are not nested in rsx
\item If a component does not have a setup function, a module server will not be created for the instance
}
}
\seealso{
Other compilation: 
\code{\link{rsx_app}()},
\code{\link{rsx_ui}()}
}
\concept{compilation}
